+++ CHANGES.rst
@@ -4,12 +4,22 @@ Change Log
 git master
 ----------
 
+New features
+''''''''''''
+
+* Added support to auto-generate Binder links for examples via ``binder``
+  config. `#244 <https://github.com/sphinx-gallery/sphinx-gallery/pull/244>`_
+
 Bug Fixes
 '''''''''
 
 * Fix __future__ imports across cells. See `#308
-  <https://github.com/sphinx-gallery/sphinx-gallery/pull/308>` for more details.
+  <https://github.com/sphinx-gallery/sphinx-gallery/pull/308>`_ for more details.
 
+Developer changes
+'''''''''''''''''
+
+* Dropped support for Sphinx <= 1.4.
 
 v0.1.13
 -------